tests/wpt/web-platform-tests/payment-request/payment-request-canmakepayment-method-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Manual tests for PaymentRequest.canMakePayment() method</title>
<link rel="help" href="https://w3c.github.io/payment-request/#canmakepayment-method">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({
  explicit_done: true,
  explicit_timeout: true,
});

const basicCard = Object.freeze({ supportedMethods: "basic-card" });
const applePay = Object.freeze({
  supportedMethods: "https://apple.com/apple-pay",
  data: {
    version: 3,
    merchantIdentifier: "merchant.com.example",
    countryCode: "US",
    merchantCapabilities: ["supports3DS"],
    supportedNetworks: ["visa"],
  }
});
const defaultMethods = Object.freeze([basicCard, applePay]);
const defaultDetails = Object.freeze({
  total: {
    label: "Total",
    amount: {
      currency: "USD",
      value: "1.00",
    },
  },
});

function testInteractiveState() {
  promise_test(async t => {
    const request = new PaymentRequest(defaultMethods, defaultDetails);
    const acceptPromise = request.show(); // Sets state to "interactive"
    const canMakePaymentPromise = request.canMakePayment();
    try {
      const result = await canMakePaymentPromise;
      assert_true(
        false,
        `canMakePaymentPromise should have thrown InvalidStateError`
      );
    } catch (err) {
      await promise_rejects(t, "InvalidStateError", canMakePaymentPromise);
    } finally {
      await request.abort();
      await promise_rejects(t, "AbortError", acceptPromise);
    }
    // The state should be "closed"
    await promise_rejects(t, "InvalidStateError", request.canMakePayment());
  }, 'If request.[[state]] is "interactive", then return a promise rejected with an "InvalidStateError" DOMException.');
}

function testClosedState() {
  promise_test(async t => {
    const request = new PaymentRequest(defaultMethods, defaultDetails);
    const acceptPromise = request.show(); // Sets state to "interactive"
    acceptPromise.catch(() => {}); // no-op, just to silence unhandled rejection in devtools.
    await request.abort(); // The state is now "closed"
    await promise_rejects(t, "InvalidStateError", request.canMakePayment());
    try {
      const result = await request.canMakePayment();
      assert_true(
        false,
        `should have thrown InvalidStateError, but instead returned "${result}"`
      );
    } catch (err) {
      assert_equals(
        err.name,
        "InvalidStateError",
        "must be an InvalidStateError."
      );
    }
  }, 'If request.[[state]] is "closed", then return a promise rejected with an "InvalidStateError" DOMException.');
}
</script>

<h2>Manual tests for PaymentRequest.canMakePayment() method</h2>
<p>
  Click on each button in sequence from top to bottom without refreshing the
  page. Each button will bring up the Payment Request UI window and then will
  close it automatically. (If a payment sheet stays open, the test has failed.)
</p>
<ol>
  <li>
    <button onclick="testInteractiveState()">
      If request.[[state]] is "interactive", then return a promise rejected with
      an "InvalidStateError" DOMException.
    </button>
  </li>
  <li>
    <button onclick="testClosedState()">
      If request.[[state]] is "closed", then return a promise rejected with an
      "InvalidStateError" DOMException.
    </button>
  </li>
  <li><button onclick="done()">Done!</button></li>
</ol>
<small>
  If you find a buggy test, please <a href="https://github.com/web-platform-tests/wpt/issues">file a bug</a>
  and tag one of the <a href="https://github.com/web-platform-tests/wpt/blob/master/payment-request/META.yml">suggested reviewers</a>.
</small>
